Make the write semantics of `caching` configurable
By default, do not complete the `Future` until the cache write has
completed. This makes the behaviour deterministic, as you have a
guarantee that the value is now present in the cache when the Future
completes.

Note that this is a change in behaviour. Until now, ScalaCache would
complete the Future as quickly as possible once the value had been
computed, and perform the cache write in the background. It is possible
to maintain this behaviour by toggling a boolean in the config.